requirements.txt
amqp==2.6.1
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.1
django-allauth==0.42.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-environ==0.4.5
django-filter==2.3.0
django-js-asset==1.2.2
django-storages==1.10
environ==1.0
filelock==3.0.12
idna==2.10
isort==5.5.1
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
Pillow==7.2.0
pipenv==2020.8.13
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pylint==2.6.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
vine==1.3.0
virtualenv==20.0.31
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
